Revert "Merge remote-tracking branch 'quoc/day_night_system' into Pokemon-Upgraded-Emerald"
This reverts commit cdb221c148f3d5d5c5bb21ab3010954566da4e25, reversing
changes made to 9a28a8dbf7fe76a175325892755b83b778a4cee5.

+++ src/naming_screen.c
@@ -1395,7 +1395,7 @@ static void NamingScreen_NoIcon(void)
 
 static void NamingScreen_CreatePlayerIcon(void)
 {
-    u16 rivalGfxId;
+    u8 rivalGfxId;
     u8 spriteId;
 
     rivalGfxId = GetRivalAvatarGraphicsIdByStateIdAndGender(PLAYER_AVATAR_STATE_NORMAL, sNamingScreen->monSpecies);
@@ -2587,3 +2587,5 @@ static const struct SpritePalette sSpritePalettes[] =
     {gNamingScreenMenu_Pal[4], PALTAG_OK_BUTTON},
     {}
 };
+
+
